+++ application-domain/src/main/java/io/github/v1serviceapplication/stay/api/StayApi.java
@@ -1,9 +1,6 @@
 package io.github.v1serviceapplication.stay.api;
 
-import io.github.v1serviceapplication.stay.api.dto.response.AdminUserInfoResponse;
-import io.github.v1serviceapplication.stay.api.dto.response.QueryAllStayStatusResponse;
-import io.github.v1serviceapplication.stay.api.dto.response.QueryStayStatusCodeResponse;
-import io.github.v1serviceapplication.stay.api.dto.response.QueryStayStatusResponse;
+import io.github.v1serviceapplication.stay.api.dto.response.*;
 import io.github.v1serviceapplication.stay.code.StayStatusCode;
 
 import java.util.UUID;
@@ -17,4 +14,6 @@ public interface StayApi {
     QueryAllStayStatusResponse queryAllStayStatus();
     AdminUserInfoResponse queryUserInfo(UUID userId);
     void adminChangeStayStatus(UUID userId, StayStatusCode status);
+
+    StayApplyListResponse queryStayApplyList();
 }
