+++ application-domain/src/main/java/io/github/v1serviceapplication/stay/api/dto/response/StayOption.java
@@ -1,4 +1,4 @@
-package io.github.v1serviceapplication.infrastructure.excel;
+package io.github.v1serviceapplication.stay.api.dto.response;
 
 public class StayOption {
     public static final String[] VALUES = new String[] {"금요귀가", "토요귀가", "토요귀사", "잔류"};
