+++ application-domain/src/main/java/io/github/v1serviceapplication/stay/api/dto/response/StayStatus.java
@@ -1,7 +1,6 @@
-package io.github.v1serviceapplication.infrastructure.excel.presentation.dto;
+package io.github.v1serviceapplication.stay.api.dto.response;
 
-import io.github.v1serviceapplication.global.error.exception.InvalidStayStatusException;
-import io.github.v1serviceapplication.infrastructure.excel.StayOption;
+import io.github.v1serviceapplication.stay.exception.InvalidStayStatusException;
 import lombok.Builder;
 import lombok.Getter;
