+++ application-domain/src/main/java/io/github/v1serviceapplication/stay/service/StayApiImpl.java
@@ -110,4 +110,9 @@ public class StayApiImpl implements StayApi {
     public void adminChangeStayStatus(UUID userId, StayStatusCode status) {
         stayRepositorySpi.changeStayStatus(userId, status);
     }
+
+    @Override
+    public StayApplyListResponse queryStayApplyList() {
+        return stayRepositorySpi.queryStayApplyList();
+    }
 }
