+++ application-infrastructure/src/main/java/io/github/v1serviceapplication/domain/stay/domain/repository/StayRepositorySpiImpl.java
@@ -2,6 +2,8 @@ package io.github.v1serviceapplication.domain.stay.domain.repository;
 
 import io.github.v1serviceapplication.domain.stay.domain.StayEntity;
 import io.github.v1serviceapplication.domain.stay.mapper.StayMapper;
+import io.github.v1serviceapplication.stay.api.dto.response.StayApplyListResponse;
+import io.github.v1serviceapplication.stay.api.dto.response.StayStatus;
 import io.github.v1serviceapplication.error.StayNotFoundException;
 import io.github.v1serviceapplication.error.UserNotFoundException;
 import io.github.v1serviceapplication.stay.Stay;
@@ -9,17 +11,21 @@ import io.github.v1serviceapplication.stay.api.dto.response.QueryStayStatusRespo
 import io.github.v1serviceapplication.stay.api.dto.response.UserStayStatusValueResponse;
 import io.github.v1serviceapplication.stay.code.StayStatusCode;
 import io.github.v1serviceapplication.stay.spi.StayRepositorySpi;
+import io.github.v1serviceapplication.studyroom.spi.StudyRoomUserFeignSpi;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
 import javax.transaction.Transactional;
 import java.util.List;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Repository
 @RequiredArgsConstructor
 public class StayRepositorySpiImpl implements StayRepositorySpi {
     private final StayRepository stayRepository;
+    private final StayMapper stayMapper;
+    private final StudyRoomUserFeignSpi studyRoomUserFeignSpi;
 
     @Override
     @Transactional
@@ -92,6 +98,43 @@ public class StayRepositorySpiImpl implements StayRepositorySpi {
                 .orElseThrow(() -> UserNotFoundException.EXCEPTION);
 
         stay.changeCode(stayStatusCode);
+    }
+
+    @Override
+    public StayApplyListResponse queryStayApplyList() {
+        return StayApplyListResponse.builder()
+                .students(getStayStatusList())
+                .build();
+    }
+
+    private List<StayStatus> getStayStatusList() {
+        List<StayStatus> stayList = stayRepository.findAll()
+                .stream()
+                .map(stayMapper::mapToStayStatus)
+                .toList();
+
+        return studyRoomUserFeignSpi.queryAllUser()
+                .stream()
+                .map(user -> {
+                    String stay = stayList.stream()
+                            .filter(stayStatus -> user.getUserId().equals(stayStatus.getUserId()))
+                            .map(StayStatus::getStay)
+                            .findFirst()
+                            .orElseGet(() -> "금요귀가");
+
+                    return StayStatus.builder()
+                            .userId(user.getUserId())
+                            .name(user.getStudentName())
+                            .num(getStudentNum(user.getGrade(), user.getClassNum(), user.getNum()))
+                            .stay(stay)
+                            .build();
+
+                })
+                .collect(Collectors.toList());
+    }
+
+    private String getStudentNum(int grade, int classNum, int num) {
+        return String.valueOf(grade) + classNum + String.format("%02d", num);
     }
 }
