+++ application-infrastructure/src/main/java/io/github/v1serviceapplication/domain/stay/mapper/StayMapperImpl.java
@@ -1,7 +1,7 @@
 package io.github.v1serviceapplication.domain.stay.mapper;
 
 import io.github.v1serviceapplication.domain.stay.domain.StayEntity;
-import io.github.v1serviceapplication.infrastructure.excel.presentation.dto.StayStatus;
+import io.github.v1serviceapplication.stay.api.dto.response.StayStatus;
 import io.github.v1serviceapplication.studyroom.api.dto.response.StudentElement;
 import io.github.v1serviceapplication.studyroom.spi.StudyRoomUserFeignSpi;
 import lombok.RequiredArgsConstructor;
