+++ application-infrastructure/src/main/java/io/github/v1serviceapplication/global/security/SecurityConfig.java
@@ -32,32 +32,19 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 
                 .authorizeRequests()
                 .requestMatchers(CorsUtils::isPreFlightRequest).permitAll()
-                .antMatchers(HttpMethod.GET, "/study-room").hasAnyRole(STUDENT, SCHOOL)
-                .antMatchers(HttpMethod.POST, "/study-room").hasAnyRole(STUDENT, SCHOOL)
-                .antMatchers(HttpMethod.DELETE, "/study-room/{study_room_id}").hasAnyRole(STUDENT, SCHOOL)
-                .antMatchers(HttpMethod.GET, "/study-room/status").hasAnyRole(STUDENT, SCHOOL)
-                .antMatchers(HttpMethod.GET, "/weekend-meal").hasAnyRole(STUDENT, SCHOOL)
-                .antMatchers(HttpMethod.POST, "/weekend-meal").hasAnyRole(STUDENT, SCHOOL)
+                .antMatchers("/study-room/**").hasAnyRole(STUDENT, SCHOOL)
+                .antMatchers( "/weekend-meal").hasAnyRole(STUDENT, SCHOOL)
                 .antMatchers(HttpMethod.POST, "/picnic").hasAnyRole(STUDENT, SCHOOL)
-                .antMatchers(HttpMethod.PUT, "/stay").hasAnyRole(STUDENT, SCHOOL)
-                .antMatchers(HttpMethod.GET, "/stay").hasAnyRole(STUDENT, SCHOOL)
+                .antMatchers("/stay").hasAnyRole(STUDENT, SCHOOL)
                 .antMatchers(HttpMethod.GET, "/stay/codes/status").hasAnyRole(STUDENT, SCHOOL)
                 .antMatchers(HttpMethod.POST, "/stay/signup").hasAnyRole(STUDENT, SCHOOL)
                 .antMatchers(HttpMethod.DELETE, "/stay/signup/{user-uuid}").hasAnyRole(STUDENT, SCHOOL)
-                .antMatchers(HttpMethod.GET, "/admin/weekend-meal").hasAnyRole(DORMITORY)
-                .antMatchers(HttpMethod.GET, "/admin/picnic").hasAnyRole(DORMITORY)
-                .antMatchers(HttpMethod.GET, "/admin/stay").hasAnyRole(DORMITORY)
+                .antMatchers("/admin/**").hasAnyRole(DORMITORY)
                 .antMatchers(HttpMethod.GET, "/weekend-meal/excel").hasAnyRole(DORMITORY)
-                .antMatchers(HttpMethod.PATCH, "/admin/picnic/arrive/{picnic-id}").hasAnyRole(DORMITORY)
-                .antMatchers(HttpMethod.PATCH, "/admin/picnic/accept/{picnic-id}").hasAnyRole(DORMITORY)
-                .antMatchers(HttpMethod.DELETE, "/admin/picnic/refuse/{picnic-id}").hasAnyRole(DORMITORY)
-                .antMatchers(HttpMethod.GET, "/admin/stay/{student-id}").hasAnyRole(DORMITORY)
-                .antMatchers(HttpMethod.PUT, "/admin/stay/{student-id}").hasAnyRole(DORMITORY)
                 .antMatchers(HttpMethod.GET, "/stay/list").hasAnyRole(DORMITORY)
                 .antMatchers(HttpMethod.GET, "/stay/excel").hasAnyRole(DORMITORY)
                 .antMatchers(HttpMethod.GET, "/weekend-meal/excel").hasAnyRole(DORMITORY)
                 .antMatchers(HttpMethod.GET, "/picnic/excel").hasAnyRole(DORMITORY)
-                .antMatchers(HttpMethod.GET, "/admin/picnic/detail/{picnic-id}").hasAnyRole(DORMITORY)
                 .antMatchers("/swagger-ui/**").permitAll()
                 .antMatchers("/docs/**").permitAll()
                 .anyRequest().authenticated()
